+++ lib/pages/settings_page.dart
@@ -1,11 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:settings_ui/settings_ui.dart';
+import 'package:package_info/package_info.dart';
 
 class SettingsPage extends StatefulWidget {
   final int quality;
   final Function changeQuality;
+  final Function changeCurrentIndex;
 
   const SettingsPage(
-      {super.key, required this.quality, required this.changeQuality});
+      {super.key,
+      required this.quality,
+      required this.changeQuality,
+      required this.changeCurrentIndex});
 
   @override
   // ignore: library_private_types_in_public_api
@@ -13,12 +19,12 @@ class SettingsPage extends StatefulWidget {
 }
 
 class _SettingsPageState extends State<SettingsPage> {
-  String? dropdownValue;
+  String? qualityValue;
 
   @override
   void initState() {
     super.initState();
-    dropdownValue = widget.quality == 0
+    qualityValue = widget.quality == 0
         ? 'Normal'
         : widget.quality == 1
             ? 'High'
@@ -27,42 +33,149 @@ class _SettingsPageState extends State<SettingsPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              const Text(
-                'Quality: ',
-                style: TextStyle(fontSize: 20),
-              ),
-              DropdownButton<String>(
-                value: dropdownValue,
-                onChanged: (String? newValue) {
-                  setState(() {
-                    dropdownValue = newValue;
-                    widget.changeQuality(newValue == 'Normal'
-                        ? 0
-                        : newValue == 'High'
-                            ? 1
-                            : 2);
-                  });
-                },
-                items: <String>['Normal', 'High', 'Ultimate']
-                    .map<DropdownMenuItem<String>>((String value) {
-                  return DropdownMenuItem<String>(
-                    value: value,
-                    child: Text(value),
-                  );
-                }).toList(),
-              ),
-            ],
+    return SettingsList(sections: [
+      SettingsSection(
+        title: const Text('Common Settings'),
+        tiles: <SettingsTile>[
+          SettingsTile.navigation(
+            leading: const Icon(Icons.language),
+            title: const Text('Language'),
+            value: const Text('English'),
           ),
-          // Add more settings widgets here
+          SettingsTile.navigation(
+            leading: const Icon(Icons.music_note),
+            title: const Text('Quality'),
+            value: Text(qualityValue!),
+            onPressed: (context) {
+              showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: const Text('Select Quality'),
+                    content: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: <Widget>[
+                        ListTile(
+                          title: const Text('Normal (16bit/44.1kHz MP3)'),
+                          onTap: () {
+                            widget.changeQuality(0);
+                            setState(() {
+                              qualityValue = 'Normal';
+                            });
+                            Navigator.pop(context);
+                          },
+                        ),
+                        ListTile(
+                          title: const Text('High (24bit/48kHz FLAC)'),
+                          onTap: () {
+                            widget.changeQuality(1);
+                            setState(() {
+                              qualityValue = 'High';
+                            });
+                            Navigator.pop(context);
+                          },
+                        ),
+                        ListTile(
+                          title: const Text('Ultimate (24bit/96kHz WAV)'),
+                          onTap: () {
+                            widget.changeQuality(2);
+                            setState(() {
+                              qualityValue = 'Ultimate';
+                            });
+                            Navigator.pop(context);
+                          },
+                        ),
+                      ],
+                    ),
+                  );
+                },
+              );
+            },
+          )
         ],
       ),
-    );
+      SettingsSection(
+        title: const Text('Version'),
+        tiles: <SettingsTile>[
+          SettingsTile(
+            title: const Text('App Version'),
+            value: FutureBuilder<PackageInfo>(
+              future: PackageInfo.fromPlatform(),
+              builder: (context, snapshot) {
+                if (snapshot.hasData) {
+                  return Text(snapshot.data!.version);
+                } else {
+                  return const Text('Loading...');
+                }
+              },
+            ),
+            leading: const Icon(Icons.info),
+          ),
+        ],
+      ),
+      SettingsSection(
+        title: const Text('Credits'),
+        tiles: <SettingsTile>[
+          SettingsTile(
+            title: const Text('Developed by'),
+            value: const Text('Thomas Dumond aka LKP'),
+            leading: const Icon(Icons.developer_mode),
+          ),
+          SettingsTile(
+            title: const Text('SKOT Members'),
+            value: const Text('LKP, Anybalsmith, Lesaul'),
+            leading: const Icon(Icons.group),
+            onPressed: (context) {
+              showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: const Text('SKOT Members'),
+                    content: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        const Text('''
+SKOT is a group of friends who are passionate about music and technology. Created in 2024 by LKP, Anybalsmith and Lesaul, the label is continously growing. The main goal of SKOT is to provide the best music experience to the world.''',
+                            textAlign: TextAlign.justify),
+                        const SizedBox(height: 10),
+                        const Align(
+                          alignment: Alignment.centerLeft,
+                          child: Text('Members:', textAlign: TextAlign.left),
+                        ),
+                        ListTile(
+                          leading: const Icon(Icons.person),
+                          title: const Text('LKP'),
+                          onTap: () {
+                            widget.changeCurrentIndex(6, username: 'LKP');
+                            Navigator.pop(context);
+                          },
+                        ),
+                        ListTile(
+                          leading: const Icon(Icons.person),
+                          title: const Text('Anybalsmith'),
+                          onTap: () {
+                            widget.changeCurrentIndex(6,
+                                username: 'Anybalsmith');
+                            Navigator.pop(context);
+                          },
+                        ),
+                        ListTile(
+                          leading: const Icon(Icons.person),
+                          title: const Text('Lesaul'),
+                          onTap: () {
+                            widget.changeCurrentIndex(6, username: 'Lesaul');
+                            Navigator.pop(context);
+                          },
+                        ),
+                      ],
+                    ),
+                  );
+                },
+              );
+            },
+          ),
+        ],
+      )
+    ]);
   }
 }